move sheet picker and service into toggle modes
