+++ src/contrib/google/sheets/SheetsFileWidget.module.scss
