+++ src/contrib/google/sheets/SheetsFileWidget.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import "./FileWidget.module.scss";
+import "./SheetsFileWidget.module.scss";
 
 import React, { useCallback, useEffect, useState } from "react";
 import { type Data, type SheetMeta } from "@/contrib/google/sheets/types";
@@ -33,18 +33,21 @@ import AsyncButton from "@/components/AsyncButton";
 import { type Expression } from "@/core";
 import { isExpression } from "@/runtime/mapArgs";
 import WorkshopMessageWidget from "@/components/fields/schemaFields/widgets/WorkshopMessageWidget";
+import { type SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
 
 const API_KEY = process.env.GOOGLE_API_KEY;
 const APP_ID = process.env.GOOGLE_APP_ID;
 
-type FileWidgetProps = {
-  id?: string;
-  name: string;
+export type SheetsFileWidgetProps = SchemaFieldProps & {
   doc: SheetMeta | null;
   onSelect: (doc: SheetMeta) => void;
 };
 
-const FileWidget: React.FC<FileWidgetProps> = ({ doc, onSelect, ...props }) => {
+const SheetsFileWidget: React.FC<SheetsFileWidgetProps> = ({
+  doc,
+  onSelect,
+  ...props
+}) => {
   const [field, , helpers] = useField<string | Expression>(props);
   const [sheetError, setSheetError] = useState(null);
 
@@ -176,4 +179,4 @@ const FileWidget: React.FC<FileWidgetProps> = ({ doc, onSelect, ...props }) => {
   );
 };
 
-export default FileWidget;
+export default SheetsFileWidget;
