+++ src/services/serviceUtils.ts
@@ -43,7 +43,10 @@ const SERVICE_ID_REGEX =
 /**
  * Return the registry ids of services supported by a JSON Schema field definition
  */
-export function extractServiceIds(schema: Schema): RegistryId[] {
+export function extractServiceIds(
+  schema: Schema,
+  suppressError?: boolean
+): RegistryId[] {
   if ("$ref" in schema) {
     const match = SERVICE_ID_REGEX.exec(schema.$ref ?? "");
     return match ? [match.groups.id as RegistryId] : [];
@@ -52,10 +55,20 @@ export function extractServiceIds(schema: Schema): RegistryId[] {
   if ("anyOf" in schema) {
     return schema.anyOf
       .filter((x) => x !== false)
-      .flatMap((x) => extractServiceIds(x as Schema));
+      .flatMap((x) => extractServiceIds(x as Schema, true));
   }
 
-  throw new Error("Expected $ref or anyOf in schema for service");
+  if ("oneOf" in schema) {
+    return schema.oneOf
+      .filter((x) => x !== false)
+      .flatMap((x) => extractServiceIds(x as Schema, true));
+  }
+
+  if (suppressError) {
+    return [];
+  }
+
+  throw new Error("Expected $ref, anyOf, or oneOf in schema for service");
 }
 
 export async function locateWithRetry(
